refactor: move use case implementation to service package
